Move options related files into options directory
Summary:
Move options related files into options directory.

Test Plan:
run all existing tests.
