+++ db/compaction.h
@@ -11,7 +11,7 @@
 #include "db/version_set.h"
 #include "util/arena.h"
 #include "util/autovector.h"
-#include "util/cf_options.h"
+#include "options/cf_options.h"
 
 namespace rocksdb {
