+++ db/compaction_picker.h
@@ -20,7 +20,7 @@
 #include "rocksdb/env.h"
 #include "rocksdb/options.h"
 #include "rocksdb/status.h"
-#include "util/cf_options.h"
+#include "options/cf_options.h"
 
 namespace rocksdb {
