+++ db/db_impl.cc
@@ -79,7 +79,7 @@
 #include "util/auto_roll_logger.h"
 #include "util/autovector.h"
 #include "util/build_version.h"
-#include "util/cf_options.h"
+#include "options/cf_options.h"
 #include "util/coding.h"
 #include "util/compression.h"
 #include "util/crc32c.h"
@@ -90,8 +90,8 @@
 #include "util/log_buffer.h"
 #include "util/logging.h"
 #include "util/mutexlock.h"
-#include "util/options_helper.h"
-#include "util/options_parser.h"
+#include "options/options_helper.h"
+#include "options/options_parser.h"
 #include "util/perf_context_imp.h"
 #include "util/sst_file_manager_impl.h"
 #include "util/stop_watch.h"
