+++ db/db_iter.h
@@ -16,7 +16,7 @@
 #include "rocksdb/iterator.h"
 #include "util/arena.h"
 #include "util/autovector.h"
-#include "util/cf_options.h"
+#include "options/cf_options.h"
 
 namespace rocksdb {
