+++ db/repair.cc
@@ -80,7 +80,7 @@
 #include "rocksdb/options.h"
 #include "rocksdb/write_buffer_manager.h"
 #include "table/scoped_arena_iterator.h"
-#include "util/cf_options.h"
+#include "options/cf_options.h"
 #include "util/file_reader_writer.h"
 #include "util/filename.h"
 #include "util/string_util.h"
