+++ db/transaction_log_impl.h
@@ -14,7 +14,7 @@
 #include "rocksdb/options.h"
 #include "rocksdb/transaction_log.h"
 #include "rocksdb/types.h"
-#include "util/db_options.h"
+#include "options/db_options.h"
 #include "util/filename.h"
 
 namespace rocksdb {
