+++ db/wal_manager.h
@@ -23,7 +23,7 @@
 #include "rocksdb/status.h"
 #include "rocksdb/transaction_log.h"
 #include "rocksdb/types.h"
-#include "util/db_options.h"
+#include "options/db_options.h"
 
 namespace rocksdb {
