+++ java/rocksjni/write_batch_test.cc
@@ -21,7 +21,7 @@
 #include "rocksdb/write_buffer_manager.h"
 #include "rocksjni/portal.h"
 #include "table/scoped_arena_iterator.h"
-#include "util/cf_options.h"
+#include "options/cf_options.h"
 #include "util/logging.h"
 #include "util/testharness.h"
