+++ options/options_helper.cc
@@ -2,7 +2,7 @@
 //  This source code is licensed under the BSD-style license found in the
 //  LICENSE file in the root directory of this source tree. An additional grant
 //  of patent rights can be found in the PATENTS file in the same directory.
-#include "util/options_helper.h"
+#include "options/options_helper.h"
 
 #include <cassert>
 #include <cctype>
