+++ options/options_test.cc
@@ -20,9 +20,9 @@
 #include "rocksdb/convenience.h"
 #include "rocksdb/memtablerep.h"
 #include "rocksdb/utilities/leveldb_options.h"
-#include "util/options_helper.h"
-#include "util/options_parser.h"
-#include "util/options_sanity_check.h"
+#include "options/options_helper.h"
+#include "options/options_parser.h"
+#include "options/options_sanity_check.h"
 #include "util/random.h"
 #include "util/stderr_logger.h"
 #include "util/testharness.h"
