+++ table/block_based_table_reader.h
@@ -27,7 +27,7 @@
 #include "table/table_properties_internal.h"
 #include "table/table_reader.h"
 #include "table/two_level_iterator.h"
-#include "util/cf_options.h"
+#include "options/cf_options.h"
 #include "util/coding.h"
 #include "util/file_reader_writer.h"
