+++ table/format.h
@@ -17,7 +17,7 @@
 
 #include "port/port.h"  // noexcept
 #include "table/persistent_cache_options.h"
-#include "util/cf_options.h"
+#include "options/cf_options.h"
 
 namespace rocksdb {
