+++ table/table_builder.h
@@ -16,7 +16,7 @@
 #include "db/table_properties_collector.h"
 #include "rocksdb/options.h"
 #include "rocksdb/table_properties.h"
-#include "util/cf_options.h"
+#include "options/cf_options.h"
 #include "util/file_reader_writer.h"
 
 namespace rocksdb {
