+++ tools/db_bench_tool_test.cc
@@ -9,7 +9,7 @@
 
 #include "rocksdb/db_bench_tool.h"
 #include "rocksdb/utilities/options_util.h"
-#include "util/options_parser.h"
+#include "options/options_parser.h"
 #include "util/random.h"
 #include "util/testharness.h"
 #include "util/testutil.h"
