+++ tools/sst_dump_tool.cc
@@ -34,7 +34,7 @@
 #include "table/meta_blocks.h"
 #include "table/plain_table_factory.h"
 #include "table/table_reader.h"
-#include "util/cf_options.h"
+#include "options/cf_options.h"
 #include "util/compression.h"
 #include "util/random.h"
