+++ util/file_util.h
@@ -9,7 +9,7 @@
 #include "rocksdb/env.h"
 #include "rocksdb/status.h"
 #include "rocksdb/types.h"
-#include "util/db_options.h"
+#include "options/db_options.h"
 
 namespace rocksdb {
 // use_fsync maps to options.use_fsync, which determines the way that
