+++ utilities/column_aware_encoding_util.h
@@ -13,7 +13,7 @@
 #include "include/rocksdb/options.h"
 #include "include/rocksdb/status.h"
 #include "table/block_based_table_reader.h"
-#include "util/cf_options.h"
+#include "options/cf_options.h"
 
 namespace rocksdb {
