+++ utilities/date_tiered/date_tiered_db_impl.h
@@ -12,7 +12,7 @@
 
 #include "rocksdb/db.h"
 #include "rocksdb/utilities/date_tiered_db.h"
-#include "util/cf_options.h"
+#include "options/cf_options.h"
 #include "util/instrumented_mutex.h"
 
 namespace rocksdb {
